test(packages): name the progress tests after the file they test
There is no packages_composite_progress.go. The progress scale these tests
cover is in packages_composite.go, so by this project's convention they belong
in that module's test file, and a reader looking for them now finds them where
the convention says to look.

No test is lost or changed: the same 150 test functions run, and the merge only
reconciles the two import blocks.
